Allow user to specify their own file extensions
The benefit now is that you can specify an exclusive list of
extensions to process, for instance only `jsx` files.

Ideally, the existing `--extension|-x` option can be removed
because you can specify as many extensions as you want now, but
I left it in there because I didn't want to break any existing
functionality.